[ui] Backfill status tag
